+++ parse/process_options.go
@@ -1,0 +1,23 @@
+package parse
+
+import "io"
+
+type options struct {
+	w      io.Writer
+	follow bool
+	debug  bool
+}
+
+type OptionsFunc func(o *options)
+
+func WithFollowOutput(b bool) OptionsFunc {
+	return func(o *options) { o.follow = b }
+}
+
+func WithWriter(w io.Writer) OptionsFunc {
+	return func(o *options) { o.w = w }
+}
+
+func WithDebug() OptionsFunc {
+	return func(o *options) { o.debug = true }
+}
